Chapter 12/backend/requirements.txt
asgiref==3.11.0
attrs==25.3.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Django==5.2.3
django-allauth==65.3.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.1
websockets==15.0.1
whitenoise==6.11.0
